% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_myfun.R
\name{all_duplicate}
\alias{all_duplicate}
\title{list duplicated data}
\usage{
all_duplicate(DT_, ...)
}
\arguments{
\item{DT_}{data_frame}

\item{...}{comapring variable}
}
\value{
data_frame. duplicated data
}
\description{
list duplicated data
}
