% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_recode.R
\name{change_posi}
\alias{change_posi}
\title{Title high efficiency recode}
\usage{
change_posi(recordtab, old = NULL, quickNA = T)
}
\arguments{
\item{recordtab}{data_frame. The table of the recode rule. V1 is old value; V2 is new value}

\item{old}{vector. recoded aims}

\item{quickNA}{efficienct option}
}
\value{
recoded vector
}
\description{
when recode rules are more than 10000, this function is more efficienct
}
