% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unique.R
\name{check_unique}
\alias{check_unique}
\title{\code{check_unique} check all vlaue is unique}
\usage{
check_unique(vec = NULL)
}
\arguments{
\item{vec}{vector}
}
\value{
vector
}
\description{
check all vlaue is unique
}
