% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coltype.R
\name{coltype}
\alias{coltype}
\title{\code{coltype} set reading datatype}
\usage{
coltype(
  len,
  at = list(),
  defult = c("?", "guess"),
  readfun = c("read_csv", "read_excel")
)
}
\arguments{
\item{len}{set variable datatype}

\item{at}{setting variable index}

\item{defult}{defult datatype. read_csv using "?", and read_excel using "guess"}

\item{readfun}{reading function, set which function do you use to read file.}
}
\value{
coltype
}
\description{
can set datatpy when using read_csv or read_excel more efficiently
}
