% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_position.R
\name{get_position}
\alias{get_position}
\title{\code{get_position} get position}
\usage{
get_position(vec, search, ret1 = 1)
}
\arguments{
\item{vec}{vector}

\item{search}{conditoin}

\item{ret1}{return all or first one, default is first one.}
}
\value{
index vector
}
\description{
get index value
}
