% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRC.R
\name{getRC}
\alias{getRC}
\title{\code{getRC} get R,C coordinate}
\usage{
getRC(matrix, condiction, weight = F, undirected = F, igraph = F)
}
\arguments{
\item{matrix}{matrix or igraph matrix}

\item{condiction}{example matrix>3}

\item{weight}{show values in matrix}

\item{undirected}{undirected}

\item{igraph}{matrix is igraph data?}
}
\value{
edgelist data_frame
}
\description{
get matrix cell R,C coordinate with specific condition
}
