% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{\code{gg_color_hue} generate ggplot colors.}
\usage{
gg_color_hue(n, change = list())
}
\arguments{
\item{n}{How many color hueps do you want?}
}
\value{
thml  color codes.
}
\description{
This function can generate html color codes from ggplot color styles.
}
\examples{
gg_color_hue(5)
#F8766D #A3A500 #00BF7D #00B0F6 #E76BF3
}
