% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_network.R
\name{isolate_posi}
\alias{isolate_posi}
\title{\code{isolate_posi} find network layout outliers}
\usage{
isolate_posi(l, a = 0.7, irq = 1.5)
}
\arguments{
\item{l}{matrix. network plot position}

\item{a}{recale time}

\item{irq}{define outliers index}
}
\value{
matrix
}
\description{
find network layout outliers and move them close center
}
