% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_readdata.R
\name{makeDT.var}
\alias{makeDT.var}
\title{\code{makeDT.var} generate varible label table}
\usage{
makeDT.var(DT = DT)
}
\arguments{
\item{DT}{data_frame}
}
\value{
varible label table
}
\description{
In spss, there are two table in one data, varible label table and varible value table. However, there is only varible value table in R, now this funcion can help you generate varible label table.
}
