% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_readdata.R
\name{makeDT_lab}
\alias{makeDT_lab}
\title{\code{makeDT.lab} read spss and dta}
\usage{
makeDT_lab(
  path = "",
  type = c("sav", "dta"),
  label_length = 200,
  reencode = NULL
)
}
\arguments{
\item{path}{file path}

\item{type}{default="sav". Read dta(stata) or sav(spss)}

\item{label_length}{default=200. how long \code{variable.labels} you want}

\item{reencode}{default="Utf-8". some data need to change into "Big5"}
}
\value{
data_frame
}
\description{
This function is based on \code{foreign} and \code{readstata13} this two package, but use this function can operater data easier after some recode dealing.
}
\details{
There are three attribute providing for user to konw variable information in each variables. One is \code{variable.labels}, which provides explanation of variable, the other is \code{value.labels},which provides value label for user to contrast value to it's meaning, and the other is \code{table} which summarises variable distribution.
}
