% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_recode.R
\name{nato}
\alias{nato}
\title{recode NA}
\usage{
nato(vec, to)

nato(vec, to)
}
\arguments{
\item{vec}{vector}

\item{to}{recode into ?}
}
\value{
vector

vector
}
\description{
recode NA into specific value

recode NA into ???
}
