% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newversion.R
\name{newversion}
\alias{newversion}
\title{Title new version data names}
\usage{
newversion(path, filename, nversion = F, update = T)
}
\arguments{
\item{path}{file path ,without filename}

\item{filename}{filename}

\item{nversion}{new version}

\item{update}{}
}
\value{

}
\description{
Title new version data names
}
