% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.table.R
\name{re.table}
\alias{re.table}
\title{\code{re.table} recode table}
\usage{
re.table(vec = vec)
}
\arguments{
\item{vec}{vector or data_frame}
}
\value{
vector or data_frame
}
\description{
recode data attribete's table
}
