% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_recode.R
\name{re_table}
\alias{re_table}
\title{\code{re.table} recode table}
\usage{
re_table(vec = vec)
}
\arguments{
\item{vec}{vector or data_frame}
}
\value{
vector or data_frame
}
\description{
recode data attribete's table
}
