% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_recode.R
\name{rec.dym}
\alias{rec.dym}
\title{\code{rec.dym} recode dummy variables}
\usage{
\method{rec}{dym}(vec = vec, DT = DT, name = "dym", nominal = F)
}
\arguments{
\item{vec}{nominal variable}

\item{DT}{Data}

\item{name}{dummy variables' name}

\item{nominal}{dependent variable of nominal regression type}
}
\value{
Data
}
\description{
recode factor into dummy variables
}
