% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_new.R
\name{rec.new}
\alias{rec.new}
\title{\code{rec.new} YC style recode}
\usage{
\method{rec}{new}(
  vec = NULL,
  rec = NULL,
  type = c("n", "f", "c", "l"),
  ref = NULL,
  level = NULL,
  labels = T,
  keep_vl = F
)
}
\arguments{
\item{vec}{vector, function can recode four type vector: numeric,factor,character,logical,}

\item{rec}{recode rule. example:"1,2,4=1;5:9=2;10=NA;else=copy"}

\item{type}{character. What new recoded vector type you want. "n":numeric. "f":factor. "c":character. "l":logical.}

\item{ref}{it works only when type=="f". You can chose which factor level being reference group.}

\item{level}{it works only when type=="f". You can rerank  factor levels order.}

\item{labels}{it works only when type=="f". Recording recode rule.}

\item{keep_vl}{default=F. Keep value.labels.}
}
\value{
vector
}
\description{
#' Same as \code{\link{rec_new}}
}
