% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_recode.R
\name{rec_dym}
\alias{rec_dym}
\title{\code{rec.dym} recode dummy variables}
\usage{
rec_dym(vec = vec, DT = DT, name = "dym", nominal = F)
}
\arguments{
\item{vec}{nominal variable}

\item{DT}{Data}

\item{name}{dummy variables' name}

\item{nominal}{dependent variable of nominal regression type}
}
\value{
Data
}
\description{
recode factor into dummy variables
}
