% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_range.R
\name{rec_range}
\alias{rec_range}
\title{rescale numeric range}
\usage{
rec_range(nuc.vec = NULL, to.min = NULL, to.max = NULL, dot = 0)
}
\arguments{
\item{nuc.vec}{an numeric vector}

\item{to.min}{rescale max value}

\item{to.max}{rescale min value}

\item{dot}{round to N'th decimal place}
}
\value{
the rescale numeric vector
}
\description{
\code{rec_range}
}
