% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_city.R
\name{recode_city}
\alias{recode_city}
\title{Urban and rural value}
\usage{
recode_city(vec, type = c(8, 4), Nato = c(NA, "copy"))
}
\arguments{
\item{vec}{postal code}

\item{type}{3 level(+ foreign) or 7 level(+ foreign)}

\item{Nato}{NA recode into ?}
}
\description{
\code{recode_city} can recode postal code into district developing level
}
