% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_ses.R
\name{recode_ses}
\alias{recode_ses}
\title{trasform isco/tsco codes}
\usage{
recode_ses(vec = NULL, from = "isco08", to = "ISEI", notchange = c(NA, "copy"))
}
\arguments{
\item{vec}{work index codes}

\item{from}{work index codes(isco,roc,tsco)}

\item{to}{codes sturcture system(isco,roc,tsco) or ses code(ISEI,NTSEI,NTOPS)}

\item{notchange}{when codes are not corresponded to rule, recode to...}
}
\value{
vector
}
\description{
\code{recode_ses} This function can trasform work index codes into ses value, and trasform from one index sturcture into the other sturcture.
}
