% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_myfun.R
\name{regpred}
\alias{regpred}
\title{\code{regpred} predict order reg}
\usage{
regpred(
  model = model,
  at = list(),
  level_name = as.character(),
  factor_list = NULL
)
}
\arguments{
\item{model}{order regression result}

\item{at}{parameter}

\item{level_name}{order level name}

\item{factor_list}{factor in regresion}
}
\value{
dataframe
}
\description{
predict marginal effect in order regression
}
