% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_max.R
\name{select_max}
\alias{select_max}
\title{\code{select_max} Find most frequence}
\usage{
select_max(vec, rank = 1, na.rm = "no", na.replace = NA, check0 = T)
}
\arguments{
\item{vec}{vector}

\item{rank}{select frequence rank, default=1}

\item{na.rm}{omits NA}

\item{na.replace}{NA recode to ?}

\item{check0}{if length==0,stop}
}
\value{
1 length vector
}
\description{
select value appearing most frequence in vector
}
\examples{
select_max(c(2,2,2,3,3,4,4,4,4,5))
#5
}
