% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_network.R
\name{twomode_onemode}
\alias{twomode_onemode}
\title{Two-mode to one-mode}
\usage{
twomode_onemode(
  edge.list = edge.list,
  type = 1,
  binary = F,
  cut.at = 1,
  twopath = F,
  outpute = c("adjacency", "edgelist1", "edgelist2", "edgelist3")
)
}
\arguments{
\item{edge.list}{two mode network(event-perosn) edgelist}

\item{type}{1 output: event-evnet ;2 output:person-person}

\item{binary}{output is 1/0}

\item{cut.at}{recode vale more than or equal to \code{cut.at} into 1,and else into 0}

\item{twopath}{1-mode network transformed from 2-mode is undirect, but it sitll can choose keep two path edge(1-2,2-1) or not(only 1-2)}

\item{outpute}{choose which datatype you want. "adjacency"=matrix, "edgelist1"=binary edgelist,"edgelist2"=edgelist with weight variable,"edgelist3"=edgelist repeat N multiplied by weight)}
}
\value{
edgelist or matrix
}
\description{
Two-mode network transform into one-mode network. This function can make 1-mode network matrix,binary edgelist, weight edgelist, and you can set it is directed or not.
}
\examples{
You can see \href{https://chang70480.github.io/YCfun/inst/html_demo/twomode_onemode.html}{here}.
}
