% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nato.R
\name{nato}
\alias{nato}
\title{recode NA}
\usage{
nato(vec, to, Inf_include = T, NaN_include = T)
}
\arguments{
\item{vec}{vector}

\item{to}{recode into ?}

\item{Inf_include}{recode including Inf?}
}
\value{
vector
}
\description{
recode NA into specific value
}
