% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_readdata.R
\name{tscs}
\alias{tscs}
\title{\code{tscs} show tscs Data}
\usage{
tscs()
}
\value{
data
}
\description{
There are many tscs R's Data in YCfun package, and they are dispersed in \code{tscsPhase1-7}.
}
\examples{
tscs()
}
